shared/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "b2b-shared"
version = "0.1.0"
description = "Pacote compartilhado entre mercado-service e negociacao-service do Portal B2B"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "python-jose[cryptography]>=3.3",
    "aiokafka>=0.10",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "orjson>=3.9",
]

[tool.setuptools.packages.find]
include = ["b2b_shared*"]
